Restructure atomic_states attribute in Projwfc
The attribute is now a list of AtomicState objects
which now have attributes that are dictionaries
with keys of Spin.up and Spin.down. This replaces
the old format of dictionaries with keys of
Spin.up and Spin.down and values of lists of
AtomicState objects. This brings the class back
in line with the DOS class and makes more sense.